requirements.txt
akshare
duckdb
matplotlib
pyyaml
requests